split database manager into own module, server::dbmgr -> core
